% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpopg-doc-parse.R
\name{cpopg_doc_parse}
\alias{cpopg_doc_parse}
\title{Parseia uma página baixada}
\usage{
cpopg_doc_parse(file_name)
}
\arguments{
\item{file_name}{vetor com os caminhos dos processos a serem parseados.}
}
\value{
tibble com os processos parseados. Cada linha é um arquivo e a coluna
output é uma list-column que contém tibbles com os resultados de cada
arquivo
}
\description{
Retorna uma tibble com os resultados. Atualmente não está pegando os
incidentes, por simplicidade. Mas depois vamos adicionar.
}
\examples{

\dontrun{
arqs <- fs::dir_ls("data-raw/cpopg/96926147868")
resultados <- cpopg_doc_parse(arqs)
}

}
